+++ src/db/fs.ts
@@ -1,9 +1,4 @@
 import initFs from '@/boot/initFs';
-import { myPath } from '@/modules/myPath';
-import {
-  ROOT_DIR_INODE_ID_KEY,
-  WORKING_DIR_INODE_ID_KEY,
-} from '@/shell/constants';
 import { DBSchema, IDBPDatabase, IDBPTransaction, openDB } from 'idb';
 
 export type Mode = 'directory' | 'symlink' | 'file';
@@ -42,288 +37,3 @@ export const fsDB = openDB<FsDBSchema>('fs', 1, {
     initFs(tx);
   },
 });
-
-export async function resolveInodeId(db: FsDB, path: string): Promise<number> {
-  const normalizedPath = myPath.normalize(path);
-  const pathSegments = normalizedPath.split('/').filter(Boolean);
-
-  const baseInodeStr = localStorage.getItem(
-    myPath.isAbsolute(normalizedPath)
-      ? ROOT_DIR_INODE_ID_KEY
-      : WORKING_DIR_INODE_ID_KEY
-  );
-  if (baseInodeStr === null) {
-    throw Error('missing inode id');
-  }
-  const baseInode = parseInt(baseInodeStr);
-  if (isNaN(baseInode)) {
-    throw Error('invalid inode id');
-  }
-
-  return resolveInodeIdRecursive(db, baseInode, pathSegments);
-}
-
-async function resolveInodeIdRecursive(
-  db: FsDB,
-  inodeId: number,
-  pathSegments: string[],
-  index = 0
-): Promise<number> {
-  if (pathSegments[index] === undefined) return inodeId;
-
-  const { directoryContents } = await readDirectory(db, inodeId);
-  const nextInodeId = directoryContents.get(pathSegments[index]);
-  if (nextInodeId === undefined) {
-    throw Error('no such file or directory');
-  }
-
-  return resolveInodeIdRecursive(db, nextInodeId, pathSegments, index + 1);
-}
-
-async function readFile(db: FsDB, directoryInodeId: number, name: string) {
-  const { directoryContents } = await readDirectory(db, directoryInodeId);
-
-  const readTx = db.transaction(['inodes', 'blobs'], 'readonly');
-  const inodes = readTx.objectStore('inodes');
-  const blobs = readTx.objectStore('blobs');
-
-  const fileInodeId = directoryContents.get(name);
-  if (fileInodeId === undefined) {
-    throw Error('file does not exist');
-  }
-
-  const fileInode = await inodes.get(fileInodeId);
-  if (!fileInode) {
-    throw Error('file inode not found');
-  }
-
-  const { blobId: fileBlobId } = fileInode;
-  const fileBlob = await blobs.get(fileBlobId);
-  if (!fileBlob) {
-    throw Error('file blob not found');
-  }
-
-  return fileBlob;
-}
-
-export async function readDirectory(db: FsDB, directoryInodeId: number) {
-  const checkDirectoryTx = db.transaction(['inodes', 'blobs'], 'readonly');
-  const inodes = checkDirectoryTx.objectStore('inodes');
-  const blobs = checkDirectoryTx.objectStore('blobs');
-
-  const directoryInode = await inodes.get(directoryInodeId);
-  if (!directoryInode) {
-    throw Error('directory inode not found');
-  }
-
-  const { blobId: directoryBlobId, modifiedTime } = directoryInode;
-  const directoryContentsBlob = await blobs.get(directoryBlobId);
-  if (!directoryContentsBlob) {
-    throw Error('directory blob not found');
-  }
-
-  const directoryContents = await deserializeDirectoryContent(
-    directoryContentsBlob
-  );
-
-  return { directoryContents, modifiedTime };
-}
-
-export async function createDirectory(
-  db: FsDB,
-  parentDirInodeId: number,
-  name: string
-) {
-  const { directoryContents: parentDirContents, modifiedTime } =
-    await readDirectory(db, parentDirInodeId);
-
-  if (parentDirContents.has(name)) {
-    throw Error('file already exists');
-  }
-
-  const writeDirTx = db.transaction(['inodes', 'blobs'], 'readwrite');
-  const inodes = writeDirTx.objectStore('inodes');
-  const blobs = writeDirTx.objectStore('blobs');
-
-  const directoryInode = await inodes.get(parentDirInodeId);
-  if (!directoryInode) {
-    throw Error('directory inode not found');
-  }
-  if (directoryInode.modifiedTime > modifiedTime) {
-    throw Error('directory inode write race condition');
-  }
-
-  const now = new Date();
-  const newDirBlobId = await blobs.add(new Blob());
-  const newDirBlob = serializeDirectoryContent(
-    new Map([
-      ['.', newDirBlobId],
-      ['..', parentDirInodeId],
-    ])
-  );
-  await blobs.put(newDirBlob, newDirBlobId);
-  const newDirInodeId = await inodes.add({
-    mode: 'directory',
-    size: 0,
-    accessedTime: now,
-    modifiedTime: now,
-    createdTime: now,
-    linkCount: 1,
-    blobId: newDirBlobId,
-  });
-
-  parentDirContents.set(name, newDirInodeId);
-  const directoryBlob = serializeDirectoryContent(parentDirContents);
-  await blobs.put(directoryBlob, directoryInode.blobId);
-  await inodes.put(
-    {
-      ...directoryInode,
-      accessedTime: now,
-      modifiedTime: now,
-    },
-    parentDirInodeId
-  );
-}
-
-export async function deleteDirectory(
-  db: FsDB,
-  parentDirInodeId: number,
-  name: string
-) {
-  const {
-    directoryContents: parentDirContents,
-    modifiedTime: parentDirModifiedTime,
-  } = await readDirectory(db, parentDirInodeId);
-
-  const targetDirInodeId = parentDirContents.get(name);
-
-  if (targetDirInodeId === undefined) {
-    throw Error('file does not exist');
-  }
-
-  const {
-    directoryContents: targetDirContents,
-    modifiedTime: targetDirModifiedTime,
-  } = await readDirectory(db, targetDirInodeId);
-
-  if (targetDirContents.size > 2) {
-    throw Error('directory not empty');
-  }
-
-  const deleteDirTx = db.transaction(['inodes', 'blobs'], 'readwrite');
-  const inodes = deleteDirTx.objectStore('inodes');
-  const blobs = deleteDirTx.objectStore('blobs');
-
-  const parentDirInode = await inodes.get(parentDirInodeId);
-  if (!parentDirInode) {
-    throw Error('parent directory inode not found');
-  }
-  if (parentDirInode.modifiedTime > parentDirModifiedTime) {
-    throw Error('parent directory inode write race condition');
-  }
-
-  const targetDirInode = await inodes.get(targetDirInodeId);
-  if (!targetDirInode) {
-    throw Error('target directory inode not found');
-  }
-  if (targetDirInode.modifiedTime > targetDirModifiedTime) {
-    throw Error('target directory inode write race condition');
-  }
-
-  await blobs.delete(targetDirInode.blobId);
-  await inodes.delete(targetDirInodeId);
-  parentDirContents.delete(name);
-
-  const now = new Date();
-  const parentDirBlob = serializeDirectoryContent(parentDirContents);
-  await blobs.put(parentDirBlob, parentDirInode.blobId);
-  await inodes.put(
-    {
-      ...parentDirInode,
-      accessedTime: now,
-      modifiedTime: now,
-    },
-    parentDirInodeId
-  );
-}
-
-export async function createFile(
-  db: FsDB,
-  directoryInodeId: number,
-  name: string,
-  content: Blob
-) {
-  const { directoryContents, modifiedTime } = await readDirectory(
-    db,
-    directoryInodeId
-  );
-
-  if (directoryContents.has(name)) {
-    throw Error('file already exists');
-  }
-
-  const writeFileTx = db.transaction(['inodes', 'blobs'], 'readwrite');
-  const inodes = writeFileTx.objectStore('inodes');
-  const blobs = writeFileTx.objectStore('blobs');
-
-  const directoryInode = await inodes.get(directoryInodeId);
-  if (!directoryInode) {
-    throw Error('directory inode not found');
-  }
-  if (directoryInode.modifiedTime > modifiedTime) {
-    throw Error('directory inode write race condition');
-  }
-
-  const now = new Date();
-  const newFileBlobId = await blobs.add(content);
-  const newFileInodeId = await inodes.add({
-    mode: 'file',
-    size: content.size,
-    accessedTime: now,
-    modifiedTime: now,
-    createdTime: now,
-    linkCount: 1,
-    blobId: newFileBlobId,
-  });
-
-  directoryContents.set(name, newFileInodeId);
-  const directoryBlob = serializeDirectoryContent(directoryContents);
-  await blobs.put(directoryBlob, directoryInode.blobId);
-  await inodes.put(
-    {
-      ...directoryInode,
-      accessedTime: now,
-      modifiedTime: now,
-    },
-    directoryInodeId
-  );
-}
-
-async function deserializeDirectoryContent(
-  content: Blob
-): Promise<Map<string, number>> {
-  const directoryString = await content.text();
-  const directoryEntries = directoryString.split('\n').map((line) => {
-    const [entry, inodeStr, rest] = line.split('\t', 2);
-    if (entry === undefined || inodeStr === undefined || rest !== undefined) {
-      throw Error('directory data corrupted');
-    }
-    const inode = parseInt(inodeStr);
-    if (isNaN(inode)) {
-      throw Error('invalid inode number');
-    }
-    return [entry, inode] as const;
-  });
-  return new Map(directoryEntries);
-}
-
-function serializeDirectoryContent(
-  directoryEntries: Map<string, number>
-): Blob {
-  const directoryContentString = directoryEntries
-    .entries()
-    .toArray()
-    .map((pair) => pair.join('\t'))
-    .join('\n');
-  return new Blob([directoryContentString]);
-}
