+++ src/shell/commands/ls.ts
@@ -1,6 +1,8 @@
-import { fsDB, readDirectory, resolveInodeId } from '@/db/fs';
+import { fsDB } from '@/db/fs';
 import { Executor } from '../executeCommand';
 import { partitionArgs } from '../utils';
+import { readDirectory } from '@/db/systemCalls/readDirectory';
+import { resolveInodeId } from '@/db/systemCalls/utils/resolveInodeId';
 
 const executor: Executor = async (argv, log) => {
   const { flags, positionals } = partitionArgs(argv);
