+++ src/shell/commands/rmdir.ts
@@ -1,7 +1,9 @@
-import { deleteDirectory, fsDB, resolveInodeId } from '@/db/fs';
+import { fsDB } from '@/db/fs';
 import { Executor } from '../executeCommand';
 import { partitionArgs } from '../utils';
 import { myPath } from '@/modules/myPath';
+import { deleteDirectory } from '@/db/systemCalls/deleteDirectory';
+import { resolveInodeId } from '@/db/systemCalls/utils/resolveInodeId';
 
 const executor: Executor = async (argv, log) => {
   const { flags, positionals } = partitionArgs(argv);
